Added support for float values in DELTA Buttons, fixed bug mentioned in #182
Former-commit-id: 02e8c94104e9fddbb21647cebdd36442b9effc27

+++ include/ViZDoomGame.h
@@ -51,9 +51,9 @@ namespace vizdoom {
         void replayEpisode(std::string filePath, unsigned int player = 0);
         bool isRunning();
 
-        void setAction(std::vector<int> const &actions);
+        void setAction(std::vector<double> const &actions);
         void advanceAction(unsigned int tics = 1, bool updateState = true);
-        double makeAction(std::vector<int> const &actions, unsigned int tics = 1);
+        double makeAction(std::vector<double> const &actions, unsigned int tics = 1);
 
         bool isNewEpisode();
         bool isEpisodeFinished();
@@ -62,7 +62,7 @@ namespace vizdoom {
         void sendGameCommand(std::string cmd);
 
         GameStatePtr getState();
-        std::vector<int> getLastAction();
+        std::vector<double> getLastAction();
 
 
         /* Buttons settings */
@@ -71,12 +71,12 @@ namespace vizdoom {
         std::vector<Button> getAvailableButtons();
         void setAvailableButtons(std::vector<Button> buttons);
 
-        void addAvailableButton(Button button, unsigned int maxValue = -1);
+        void addAvailableButton(Button button, double maxValue = -1);
         void clearAvailableButtons();
         size_t getAvailableButtonsSize();
 
-        void setButtonMaxValue(Button button, unsigned int maxValue);
-        int getButtonMaxValue(Button button);
+        void setButtonMaxValue(Button button, double maxValue);
+        double getButtonMaxValue(Button button);
 
 
         /* GameVariables getters and setters */
@@ -195,7 +195,7 @@ namespace vizdoom {
 
         std::vector<GameVariable> availableGameVariables;
         std::vector<Button> availableButtons;
-        std::vector<int> lastAction;
+        std::vector<double> lastAction;
 
         unsigned int nextStateNumber;
         unsigned int lastMapTic;
